bip32/bip37: fix signed-overflow UB in big-endian word assembly
Building a uint32 from bytes via (b[0] << 24) promotes the byte to
signed int; when b[0] >= 0x80 the shift overflows INT_MAX -> UB,
flagged by UBSAN in dogecoin_hdnode_private_ckd. Benign on two's-
complement targets (wraps to the intended value) but undefined.

Cast operands to uint32_t before shifting. Output unchanged: verified
value-identical across all 256 top-byte values, UBSAN clean.

+++ src/bip37.c
@@ -40,7 +40,7 @@ static uint32_t bip37_murmur3(const uint8_t* key, size_t len, uint32_t seed)
     uint32_t h = seed;
     size_t i = 0;
     for (; i + 4 <= len; i += 4) {
-        uint32_t k = key[i] | (key[i + 1] << 8) | (key[i + 2] << 16) | (key[i + 3] << 24);
+        uint32_t k = (uint32_t)key[i] | ((uint32_t)key[i + 1] << 8) | ((uint32_t)key[i + 2] << 16) | ((uint32_t)key[i + 3] << 24);
         k *= 0xcc9e2d51u;
         k = (k << 15) | (k >> 17);
         k *= 0x1b873593u;
